pyproject.toml
[project]
name = "code-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0.post1",
    "groq>=0.30.0",
    "huggingface-hub>=0.34.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "numpy>=2.2.6",
    "pypdf>=5.9.0",
    "pypdf2>=3.0.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "torch>=2.7.1",
]
